Tenmo App/java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Accounts.java
package com.techelevator.tenmo.model;

import java.math.BigDecimal;

public class Accounts {

    private int accountID;
    private int userID;
    private BigDecimal balance;

    public Accounts(int accountID, int userID, BigDecimal balance) {
        this.accountID = accountID;
        this.userID = userID;
        this.balance = balance;
    }

    public Accounts() {
    }

    public int getAccountID() {
        return accountID;
    }

    public void setAccountID(int accountID) {
        this.accountID = accountID;
    }

    public int getUserID() {
        return userID;
    }

    public void setUserID(int userID) {
        this.userID = userID;
    }

    public BigDecimal getBalance() {
        return balance;
    }

    public void setBalance(BigDecimal balance) {
        this.balance = balance;
    }


}
